Refactor: Split config package
