+++ internal/config/sort_parser.go
@@ -1,0 +1,33 @@
+package config
+
+import (
+	"fmt"
+	"qp/internal/consts"
+	"strings"
+)
+
+func parseSortOption(sortInput string) (SortOption, error) {
+	parts := strings.Split(sortInput, ":")
+	fieldKey := strings.ToLower(parts[0])
+	fieldType, exists := consts.FieldTypeLookup[fieldKey]
+	if !exists {
+		return SortOption{}, fmt.Errorf("invalid sort field: %s", fieldKey)
+	}
+
+	asc := true
+	if len(parts) > 1 {
+		switch parts[1] {
+		case "desc":
+			asc = false
+		case "asc":
+			asc = true
+		default:
+			return SortOption{}, fmt.Errorf("invalid sort direction: %s", parts[1])
+		}
+	}
+
+	return SortOption{
+		Field: fieldType,
+		Asc:   asc,
+	}, nil
+}
